pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bililive-utility"
version = "2.0.0"
authors = [
  { name="GamerNoTitle", email="your-email@example.com" },
]
description = "A FastAPI server for BiliLive Utility"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx[http2]",
    "platformdirs",
    "gitpython>=3.1.45",
    "qrcode>=7.4.2",
    "pillow>=10.4.0",
    "pywebview>=5.4",
    "pyinstaller>=6.14.2",
]

[project.scripts]
bililive-utility = "bililive_utility.__main__:main"
